7.Statistics feat: add button
