+++ rslang/src/components/App/App.tsx
@@ -17,24 +17,7 @@ import { SignUp } from '../../pages/SignUp';
 import { Layout } from '../Layout/Layout';
 import { WordsPage } from '../../pages/WordsPage';
 
-// import data from '../../pages/db.json';
-
-// const data = {
-//   statistics: {
-//     sprint: [
-//       {
-//         date: 34534534534554,
-//         new: 6,
-//         quantity: 3,
-//         wrong: 6,
-//         correct: 12,
-//       },
-//     ],
-//   },
-// };
-
 function App(): JSX.Element {
-
   return (
     <div className='App'>
       <Routes>
@@ -43,14 +26,7 @@ function App(): JSX.Element {
           <Route path='about' element={<AboutPage />} />
           <Route path='textbook' element={<TextbookPage />} />
           <Route path='textbook/words/group=:group&page=:page' element={<WordsPage />} />
-          <Route
-            path='statistics'
-            element={
-              <StatisticsPage
-              // date={data.statistics.sprint}
-              />
-            }
-          />
+          <Route path='statistics' element={<StatisticsPage />} />
 
           <Route path='sprint' element={<SprintPage />} />
           <Route path='audiocall' element={<AudioCallPage />} />
