+++ rslang/src/pages/PagesStyle.module.scss
@@ -1,0 +1,4 @@
+.statisticsPage {
+    display: flex;
+    justify-content: center;
+}
